+++ index.html
@@ -5,8 +5,7 @@
 
   <link rel="stylesheet" type="text/css" href="dist/css/mumuki-styles.css">
 
-  <script type="text/javascript" src="src/bower_components/jquery/dist/jquery.js"></script>
-  <script type="text/javascript" src="src/javascripts/mumuki-styles.js"></script>
+  <script type="text/javascript" src="dist/javascripts/mumuki-styles.js"></script>
 </head>
 <body>
