+++ src/javascripts/index.js
@@ -1,0 +1,8 @@
+module.exports = [
+
+  'load.js',
+  'file-browser.js',
+  'er-diagram.js',
+
+
+].map((file) => `src/javascripts/_${file}`);
